Fix binge map upon entering a binge
Upon entering a new binge, the question slug page re-renders. This sets the binge map width to 0 temporarily. However, the binge map nodes are miscalculated at this instant, with x values of 0. This makes the nodes lean to the left and the nodes stay like that when the width of the binge map is fixed as the render finishes.

To fix the issue, the node positions are not calculated if we have an invalid width for the binge map card.